requirements.txt
discord.py
honcho
typer
motor
